requirements.txt
fastapi
uvicorn
sentence-transformers
python-jose
PyJWT
requests
dotenv
cryptography
python-multipart